% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_fct.R
\name{fct_count}
\alias{fct_count}
\title{Reorder Factor Levels by Level Count}
\usage{
fct_count(factor_vec, decreasing = TRUE)
}
\arguments{
\item{factor_vec}{A factor vector whose levels are to be reordered.}

\item{decreasing}{Logical. Should the ordering be decreasing by count? Default is \code{TRUE}.}
}
\value{
A factor vector with levels reordered based on their count.
}
\description{
Reorders the levels of a factor vector based on the count of each level in the data.
}
\examples{
# Example factor vector
factor_vec <- factor(c('apple', 'banana', 'apple', 'cherry', 'banana', 'banana', 'date'))

# Reorder levels by decreasing count
fct_count(factor_vec)

# Reorder levels by increasing count
fct_count(factor_vec, decreasing = FALSE)
}
\author{
Kai Guo
}
