% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_fct.R
\name{fct_freq}
\alias{fct_freq}
\title{Reorder Factor Levels Based on Total Character Frequency}
\usage{
fct_freq(factor_vec, case = FALSE, decreasing = TRUE)
}
\arguments{
\item{factor_vec}{A factor vector whose levels are to be reordered.}

\item{case}{Logical. Should the character frequency count be case-sensitive? Default is \code{FALSE}.}

\item{decreasing}{Logical. Should the ordering be decreasing by total character frequency? Default is \code{TRUE}.}
}
\value{
A factor vector with levels reordered based on the total frequency of their constituent characters in the data.
}
\description{
Reorders the levels of a factor vector based on the total frequency of characters appearing in the vector.
}
\examples{
# Example factor vector
factor_vec <- factor(c('apple', 'banana', 'cherry', 'date', 'banana', 'apple', 'fig'))

# Reorder levels based on total character frequency
fct_freq(factor_vec)

# Reorder levels, case-sensitive
factor_vec_case <- factor(c('Apple', 'banana', 'Cherry', 'date', 'banana', 'apple', 'Fig'))
fct_freq(factor_vec_case, case = TRUE)
}
\author{
Kai Guo
}
