% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_misc.R
\name{fct_len}
\alias{fct_len}
\title{Sort Factor Levels Based on Their Length}
\usage{
fct_len(factor_vec, decreasing = FALSE)
}
\arguments{
\item{factor_vec}{A factor vector to be sorted.}

\item{decreasing}{Logical. Should the ordering be decreasing by length? Default is \code{FALSE}.}
}
\value{
A factor vector with levels reordered based on their length.
}
\description{
Reorders the levels of a factor vector based on the character length of each level.
}
\examples{
# Example factor vector
factor_vec <- factor(c('apple', 'banana', 'cherry', 'date'))

# Sort levels by length
fct_len(factor_vec)
}
\author{
Kai Guo
}
