% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_misc.R
\name{fct_level_lengths}
\alias{fct_level_lengths}
\title{Get Character Lengths of Factor Levels}
\usage{
fct_level_lengths(factor_vec)
}
\arguments{
\item{factor_vec}{A factor vector.}
}
\value{
A named numeric vector with the length of each level.
}
\description{
Calculates the number of characters in each level of a factor vector.
}
\examples{
# Example factor vector
factor_vec <- factor(c('apple', 'banana', 'cherry'))

# Get level lengths
fct_level_lengths(factor_vec)
}
\author{
Kai Guo
}
