% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_misc.R
\name{fct_pad_levels}
\alias{fct_pad_levels}
\title{Pad Factor Levels to Uniform Length}
\usage{
fct_pad_levels(factor_vec, width, pad_char = "0")
}
\arguments{
\item{factor_vec}{A factor vector.}

\item{width}{An integer specifying the desired width.}

\item{pad_char}{A character to use for padding. Default is \code{'0'}.}
}
\value{
A factor vector with padded levels.
}
\description{
Pads the levels of a factor vector with leading characters to achieve a specified width.
}
\examples{
# Example factor vector
factor_vec <- factor(c('1', '12', '123'))

# Pad levels to width 4
fct_pad_levels(factor_vec, width = 4)
}
\author{
Kai Guo
}
