% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_fct.R
\name{fct_pos}
\alias{fct_pos}
\title{Reorder Factor Levels Based on Characters at Specific Positions}
\usage{
fct_pos(factor_vec, positions, case = FALSE, decreasing = FALSE)
}
\arguments{
\item{factor_vec}{A factor vector whose levels are to be reordered.}

\item{positions}{A vector of positive integers specifying the character positions to use for reordering. Positions can be non-consecutive.}

\item{case}{Logical. Should the character comparison be case-sensitive? Default is \code{FALSE}.}

\item{decreasing}{Logical. Should the ordering be decreasing? Default is \code{FALSE}.}
}
\value{
A factor vector with levels reordered based on the specified character positions.
}
\description{
Reorders the levels of a factor vector based on the characters at specified positions within the factor levels.
}
\examples{
# Example factor vector
factor_vec <- factor(c('Apple', 'banana', 'Cherry', 'date', 'Fig', 'grape'))

# Reorder based on positions 1 and 3, case-insensitive
fct_pos(factor_vec, positions = c(1, 3))

# Reorder in decreasing order, case-sensitive
fct_pos(factor_vec, positions = 1:2, case = TRUE, decreasing = TRUE)
}
\author{
Kai Guo
}
