% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_fct.R
\name{fct_regex_freq}
\alias{fct_regex_freq}
\title{Reorder Factor Levels Based on Regex Pattern Frequency}
\usage{
fct_regex_freq(factor_vec, pattern, case = FALSE, decreasing = TRUE)
}
\arguments{
\item{factor_vec}{A factor vector whose levels are to be reordered.}

\item{pattern}{A string representing the regular expression pattern to match.}

\item{case}{Logical. Should the pattern matching be case-sensitive? Default is \code{FALSE}.}

\item{decreasing}{Logical. Should the ordering be decreasing by frequency? Default is \code{TRUE}.}
}
\value{
A factor vector with levels reordered based on the frequency of matched substrings.
}
\description{
Reorders the levels of a factor vector based on the frequency of substrings matching a regular expression.
}
\examples{
# Example factor vector
factor_vec <- factor(c('apple', 'banana', 'apricot', 'cherry', 'blueberry', 'blackberry', 'date'))

# Reorder based on pattern matching 'a'
fct_regex_freq(factor_vec, pattern = 'a')

# Reorder with case-sensitive matching
fct_regex_freq(factor_vec, pattern = '^[A-Z]', case = TRUE)
}
\author{
Kai Guo
}
