% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_misc.R
\name{fct_reverse}
\alias{fct_reverse}
\title{Reverse Factor Levels}
\usage{
fct_reverse(factor_vec)
}
\arguments{
\item{factor_vec}{A factor vector whose levels will be reversed.}
}
\value{
A factor vector with levels in reversed order.
}
\description{
Reverses the order of the levels in a factor vector.
}
\examples{
# Example factor vector
factor_vec <- factor(c('low', 'medium', 'high'))

# Reverse the levels
fct_reverse(factor_vec)
}
\author{
Kai Guo
}
