% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_fct.R
\name{fct_sort}
\alias{fct_sort}
\title{Sort Factor Levels Based on Another Vector or Column}
\usage{
fct_sort(factor_vec, by, decreasing = FALSE, na_last = TRUE)
}
\arguments{
\item{factor_vec}{A factor vector whose levels are to be sorted.}

\item{by}{A vector or data frame column used as the basis for sorting. Must be the same length as `factor_vec`.}

\item{decreasing}{Logical. Should the sorting be in decreasing order? Default is \code{FALSE}.}

\item{na_last}{Logical. Should `NA` values be put last? Default is \code{TRUE}.}
}
\value{
A factor vector with levels sorted based on `by`.
}
\description{
Sorts the levels of a factor vector based on the values of another vector or a column from a data frame. Handles cases where the sorting vector may contain `NA`s.
}
\examples{
# Example using a vector
factor_vec <- factor(c('apple', 'banana', 'cherry', 'date'))
by_vec <- c(2, 3, 1, NA)
fct_sort(factor_vec, by = by_vec)

# Example using a data frame column
data <- data.frame(
  Category = factor(c('apple', 'banana', 'cherry', 'date')),
  Value = c(2, 3, 1, NA)
)
fct_sort(data$Category, by = data$Value)
}
\author{
Kai Guo
}
