% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_split.R
\name{fct_split}
\alias{fct_split}
\title{Split Factor Levels and Reorder Based on Specified Criteria}
\usage{
fct_split(
  factor_vec,
  split_pattern,
  use_pattern = NULL,
  part = 1,
  position = NULL,
  char_freq = FALSE,
  decreasing = FALSE
)
}
\arguments{
\item{factor_vec}{A factor vector to be processed.}

\item{split_pattern}{A character vector specifying the pattern(s) or position(s) to use for splitting. Can be regular expressions or integer positions.}

\item{use_pattern}{An integer specifying which pattern to use if multiple patterns are provided. Default is \code{NULL} by using all pattern.}

\item{part}{An integer or integer vector specifying which part(s) to use after splitting (e.g., 1 for the first part). Can be a range or specific indices.}

\item{position}{An integer or integer vector specifying the character positions within the part(s) to consider.}

\item{char_freq}{Logical. Should the sorting be based on character frequencies within the specified part(s)? Default is \code{FALSE}.}

\item{decreasing}{Logical. Should the ordering be decreasing? Default is \code{FALSE}.}
}
\value{
A factor vector with levels reordered based on the specified conditions.
}
\description{
Splits the levels of a factor vector using specified patterns or positions and reorders based on specified parts or criteria.
}
\examples{
# Example factor vector with patterns
factor_vec <- factor(c('item1-sub1', 'item2_sub2', 'item3|sub3', 'item1-sub4'))

# Split by patterns '-', '_', or '|' and reorder based on the first part
fct_split(factor_vec, split_pattern = c('-', '_', '\\\\|'), part = 1)

# Use the second pattern '_' for splitting
fct_split(factor_vec, split_pattern = c('-', '_', '\\\\|'), use_pattern = 2, part = 2)

# Reorder based on character frequencies in the specified part
fct_split(factor_vec, split_pattern = '-', part = 2, char_freq = TRUE)
}
\author{
Kai Guo
}
