% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_fct.R
\name{fct_substr_freq}
\alias{fct_substr_freq}
\title{Reorder Factor Levels Based on Substring Frequency}
\usage{
fct_substr_freq(
  factor_vec,
  start_pos = NULL,
  end_pos = NULL,
  case = FALSE,
  decreasing = TRUE
)
}
\arguments{
\item{factor_vec}{A factor vector whose levels are to be reordered.}

\item{start_pos}{Positive integer. The starting position of the substring. If \code{NULL}, starts from the beginning.}

\item{end_pos}{Positive integer. The ending position of the substring. If \code{NULL}, goes to the end of the string.}

\item{case}{Logical. Should the substring comparison be case-sensitive? Default is \code{FALSE}.}

\item{decreasing}{Logical. Should the ordering be decreasing by frequency? Default is \code{TRUE}.}
}
\value{
A factor vector with levels reordered based on the frequency of substrings.
}
\description{
Reorders the levels of a factor vector based on the frequency of substrings extracted from the data.
}
\examples{
# Example factor vector with multi-byte characters
factor_vec <- factor(c('苹果', '香蕉', '苹果', '樱桃', '香蕉', '香蕉', '日期'))

# Reorder based on substrings from position 1 to 1
fct_substr_freq(factor_vec, start_pos = 1, end_pos = 1)

# Reorder from position 2 to end
fct_substr_freq(factor_vec, start_pos = 2)
factor_vec <- factor(c('apple', 'banana', 'apricot', 'cherry', 'banana', 'banana', 'date'))
fct_substr_freq(factor_vec, start_pos = 2, end_pos=3)
}
\author{
Kai Guo
}
