% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_misc.R
\name{fct_union}
\alias{fct_union}
\title{Get Union of Factor Levels from Multiple Vectors}
\usage{
fct_union(...)
}
\arguments{
\item{...}{Factor vectors to be united.}
}
\value{
A factor vector containing all unique levels from all provided factors.
}
\description{
Combines multiple factor vectors and returns a factor vector containing all unique levels.
}
\examples{
# Example factor vectors
factor_vec1 <- factor(c('apple', 'banana'))
factor_vec2 <- factor(c('banana', 'cherry'))
factor_vec3 <- factor(c('date', 'fig'))

# Get union of levels
fct_union(factor_vec1, factor_vec2, factor_vec3)
}
\author{
Kai Guo
}
