% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_misc.R
\name{ft_dummy}
\alias{ft_dummy}
\title{Create Dummy Variables from Factor Levels}
\usage{
ft_dummy(factor_vec)
}
\arguments{
\item{factor_vec}{A factor vector.}
}
\value{
A data frame where each column represents a level of the factor, containing 1s and 0s.
}
\description{
Generates a data frame of dummy variables (one-hot encoded) from a factor vector.
}
\examples{
# Example factor vector
factor_vec <- factor(c('apple', 'banana', 'apple', 'cherry'))

# Create dummy variables
ft_dummy(factor_vec)
}
\author{
Kai Guo
}
