% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_misc.R
\name{ft_duplicates}
\alias{ft_duplicates}
\title{Flag Duplicate Factor Levels}
\usage{
ft_duplicates(factor_vec)
}
\arguments{
\item{factor_vec}{A factor vector.}
}
\value{
A logical vector where \code{TRUE} indicates a duplicate level.
}
\description{
Identifies duplicate levels in a factor vector and returns a logical vector indicating which elements are duplicates.
}
\examples{
# Example factor vector
factor_vec <- factor(c('apple', 'banana', 'apple', 'cherry', 'banana'))

# Flag duplicates
ft_duplicates(factor_vec)
}
\author{
Kai Guo
}
