% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_misc.R
\name{ft_level_order}
\alias{ft_level_order}
\title{Get Order of Factor Levels in Data}
\usage{
ft_level_order(factor_vec)
}
\arguments{
\item{factor_vec}{A factor vector.}
}
\value{
A numeric vector representing the order of levels.
}
\description{
Returns a vector indicating the order in which factor levels appear in the data.
}
\examples{
# Example factor vector
factor_vec <- factor(c('banana', 'apple', 'cherry', 'apple', 'banana'))

# Get level order
ft_level_order(factor_vec)
}
\author{
Kai Guo
}
