% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_misc.R
\name{ft_sample_levels}
\alias{ft_sample_levels}
\title{Sample Levels from a Factor Vector}
\usage{
ft_sample_levels(factor_vec, size, seed = NULL)
}
\arguments{
\item{factor_vec}{A factor vector.}

\item{size}{An integer specifying the number of levels to sample.}

\item{seed}{An optional integer for setting the random seed.}
}
\value{
A factor vector containing only the sampled levels.
}
\description{
Randomly selects a specified number of levels from a factor vector.
}
\examples{
# Example factor vector
factor_vec <- factor(letters[1:10])

# Sample 5 levels
ft_sample_levels(factor_vec, size = 5, seed = 123)
}
\author{
Kai Guo
}
