% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_fct.R
\name{fct_sort_custom}
\alias{fct_sort_custom}
\title{Sort Factor Levels Using a Custom Function}
\usage{
fct_sort_custom(factor_vec, sort_func, decreasing = FALSE)
}
\arguments{
\item{factor_vec}{A factor vector to sort.}

\item{sort_func}{A function that takes a character vector (the levels) and returns a vector of the same length to sort by.}

\item{decreasing}{Logical. Should the sort be decreasing? Default is \code{FALSE}.}
}
\value{
A factor vector with levels reordered according to the custom function.
}
\description{
Reorders the levels of a factor vector based on a custom function applied to each level.
}
\examples{
# Example factor vector
factor_vec <- factor(c('apple', 'banana', 'cherry'))

# Sort levels by reverse alphabetical order
fct_sort_custom(factor_vec, function(x) -rank(x))

# Sort levels by length of the level name
fct_sort_custom(factor_vec, function(x) nchar(x))
}
\author{
Kai Guo
}
